Linear Regression Model implementation
I implement Linear Regression Model to the E-commerce dataset that is about the customers information who buys clothes online. I have four independent variable and one target variable. The business challenge is that they want to know they should focus on Mobile App or Website. I concluded that the "Time on App" variable is more important variable than "Time on Website" and it has a strong correlation with the target variable that is called "Yearly Amount Spent".