Logistic Regression Model Implementation
I implement Logistic Regression Model to the breast cancer dataset that has the information of people who has cancer. First of all I split the dataset to training and test dataset. I load the Logistic Regression Model and fit it to the data and in the end the accuracy metric of this model is 0.98 which means the difference between the y_pred and the actual y is very little and the model is a good model.