Classification with KNN
I implement K-Nearst Neighbors classifier to the diabetes dataset. The goal is to predict whether the person has diabetes or not. The dataset has 9 columns , the columns is about Glucose, Blood Pressure, Insulin,age,....and the last column is the target variable . I use the square root of y_test to determine the value of K and it is better to use the odd value for K to avoid confusion between two classes, I get the f1-score 0.65 and the accuracy-score 0.8, which means the KNN classifier performs fair good.